requirements.txt
PyQt5
pillow
pandas
openpyxl